+++ flutter_inappwebview/lib/src/in_app_webview/in_app_webview.dart
@@ -275,6 +275,8 @@ class InAppWebView extends StatefulWidget {
       MediaCaptureState? newState,
     )?
     onMicrophoneCaptureStateChanged,
+    void Function(InAppWebViewController controller, bool active)?
+    onWritingToolsActiveChanged,
     void Function(
       InAppWebViewController controller,
       Size oldContentSize,
@@ -522,6 +524,10 @@ class InAppWebView extends StatefulWidget {
                        oldState,
                        newState,
                      )
+               : null,
+           onWritingToolsActiveChanged: onWritingToolsActiveChanged != null
+               ? (controller, active) =>
+                     onWritingToolsActiveChanged.call(controller, active)
                : null,
            onContentSizeChanged: onContentSizeChanged != null
                ? (controller, oldContentSize, newContentSize) =>