+++ flutter_inappwebview_platform_interface/lib/src/in_app_browser/platform_in_app_browser.g.dart
@@ -1273,6 +1273,20 @@ enum PlatformInAppBrowserEventsMethod {
   ///{@endtemplate}
   onWindowFocus,
 
+  ///Can be used to check if the [PlatformInAppBrowserEvents.onWritingToolsActiveChanged] method is supported at runtime.
+  ///
+  ///{@template flutter_inappwebview_platform_interface.PlatformInAppBrowserEvents.onWritingToolsActiveChanged.supported_platforms}
+  ///
+  ///**Officially Supported Platforms/Implementations**:
+  ///- iOS WKWebView 18.0+ ([Official API - WKWebView.writingToolsActive](https://developer.apple.com/documentation/webkit/wkwebview/4433211-writingtoolsactive))
+  ///
+  ///**Parameters - Officially Supported Platforms/Implementations**:
+  ///- [active]: all platforms
+  ///
+  ///Use the [PlatformInAppBrowserEvents.isMethodSupported] method to check if this method is supported at runtime.
+  ///{@endtemplate}
+  onWritingToolsActiveChanged,
+
   ///Can be used to check if the [PlatformInAppBrowserEvents.onZoomScaleChanged] method is supported at runtime.
   ///
   ///{@template flutter_inappwebview_platform_interface.PlatformInAppBrowserEvents.onZoomScaleChanged.supported_platforms}
@@ -1653,6 +1667,9 @@ extension _PlatformInAppBrowserEventsMethodSupported
               TargetPlatform.android,
               TargetPlatform.iOS,
             ].contains(platform ?? defaultTargetPlatform);
+      case PlatformInAppBrowserEventsMethod.onWritingToolsActiveChanged:
+        return ((kIsWeb && platform != null) || !kIsWeb) &&
+            [TargetPlatform.iOS].contains(platform ?? defaultTargetPlatform);
       case PlatformInAppBrowserEventsMethod.onZoomScaleChanged:
         return ((kIsWeb && platform != null) || !kIsWeb) &&
             [