+++ flutter_inappwebview_platform_interface/lib/src/in_app_webview/platform_headless_in_app_webview.dart
@@ -83,6 +83,7 @@ class PlatformHeadlessInAppWebViewCreationParams
     super.shouldAllowDeprecatedTLS,
     super.onCameraCaptureStateChanged,
     super.onMicrophoneCaptureStateChanged,
+    super.onWritingToolsActiveChanged,
     super.onContentSizeChanged,
     super.onShowFileChooser,
     super.onInsertInputSuggestion,