+++ flutter_inappwebview_platform_interface/lib/src/in_app_webview/platform_inappwebview_widget.dart
@@ -91,6 +91,7 @@ class PlatformInAppWebViewWidgetCreationParams
     super.shouldAllowDeprecatedTLS,
     super.onCameraCaptureStateChanged,
     super.onMicrophoneCaptureStateChanged,
+    super.onWritingToolsActiveChanged,
     super.onContentSizeChanged,
     super.onShowFileChooser,
     super.onInsertInputSuggestion,